[Refactor] AbstractAuthenticationInterceptor로 추상화
- SessionAuthenticationInterceptor
- TokenAuthenticationInterceptor
