+++ src/main/java/nextstep/auth/authentication/AbstractAuthenticationInterceptor.java
@@ -1,0 +1,49 @@
+package nextstep.auth.authentication;
+
+import nextstep.auth.context.Authentication;
+import nextstep.member.application.CustomUserDetailsService;
+import nextstep.member.domain.LoginMember;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+public abstract class AbstractAuthenticationInterceptor implements HandlerInterceptor {
+    private final CustomUserDetailsService customUserDetailsService;
+    private final AuthenticationConverter authenticationConverter;
+
+    public AbstractAuthenticationInterceptor(CustomUserDetailsService customUserDetailsService, AuthenticationConverter authenticationConverter) {
+        this.customUserDetailsService = customUserDetailsService;
+        this.authenticationConverter = authenticationConverter;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
+        AuthenticationToken authenticationToken = authenticationConverter.convert(request);
+        Authentication authentication = authenticate(authenticationToken);
+        afterAuthentication(request, response ,authentication);
+
+        return false;
+    }
+
+    public Authentication authenticate(AuthenticationToken authenticationToken) {
+        String principal = authenticationToken.getPrincipal();
+        LoginMember userDetails = customUserDetailsService.loadUserByUsername(principal);
+        checkAuthentication(userDetails, authenticationToken);
+
+        return new Authentication(userDetails);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new AuthenticationException();
+        }
+
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new AuthenticationException();
+        }
+    }
+
+    public abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+}
